[Fix] clear filters issue
